requirements.txt
biopython==1.85
deepspeed==0.17.0
dm-tree==0.1.9
ml-collections==1.1.0
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
typing-extensions==4.14.0
pytorch_lightning==2.5.1.post0
wandb==0.19.11
git+https://github.com/NVIDIA/dllogger.git@0478734ff7be75adde8d160e04872664d1c62e5f
einops==0.8.1
tqdm==4.67.1
numpy==2.0.2
modelcif==0.7
torch==2.6.0
matplotlib==3.10.0
scikit-learn==1.6.1
hydra-core==1.3.2
colorama==0.4.6